Splits components into separate files
