+++ src/js/components/el1.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const el1 = React.createElement("h1",
+    {className: "title"},
+    "This is a component, created by React.createElement"
+);
+
+export default el1;
+
