+++ src/js/components/el2.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+var el2 =  (name) => (<p>Hello {name}! - functional component</p>);
+
+export default el2;
